+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
-use App\Http\Controllers\PassportController;
+
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use Laravel\Passport\Passport;
@@ -16,8 +17,14 @@ use Laravel\Passport\Passport;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+
+Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::post('register', [PassportController::class, 'register']);
+Route::post('register', [UserController::class, 'register']);
+Route::post('login', [UserController::class, 'login']);
+
+// Route::middleware('auth:api')->group(function () {
+    //aqui incluiremos las rutas que queramos proteger con token de acceso
+// });
